+++ src/main/java/com/telepathicgrunt/worldblender/features/WBFeatures.java
@@ -12,11 +12,7 @@ public class WBFeatures
 {
 	public static final DeferredRegister<Feature<?>> FEATURES = DeferredRegister.create(ForgeRegistries.FEATURES, WorldBlender.MODID);
 	
-    public static final RegistryObject<Feature<NoFeatureConfig>> WB_PORTAL_ALTAR = createFeature("portal_altar", WBPortalAltar::new);
-    public static final RegistryObject<Feature<NoFeatureConfig>> ANTI_FLOATING_BLOCKS_AND_SEPARATE_LIQUIDS = createFeature("anti_floating_blocks_and_separate_liquids", AntiFloatingBlocksAndSeparateLiquids::new);
-
-    private static <F extends Feature<?>> RegistryObject<F> createFeature(String name, Supplier<F> feature)
-    {
-		return FEATURES.register(name, feature);
-	}
+    public static final RegistryObject<Feature<NoFeatureConfig>> WB_PORTAL_ALTAR = FEATURES.register("portal_altar", WBPortalAltar::new);
+    public static final RegistryObject<Feature<NoFeatureConfig>> ANTI_FLOATING_BLOCKS_AND_SEPARATE_LIQUIDS = FEATURES.register("anti_floating_blocks_and_separate_liquids", AntiFloatingBlocksAndSeparateLiquids::new);
+    public static final RegistryObject<Feature<NoFeatureConfig>> ITEM_CLEARING = FEATURES.register("item_clearing", ItemClearingFeature::new);
 }